// src/api/huuray/models/CatalogueRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * An object containing the required input for the product catalogue API
 */
export type CatalogueRequest = {
    /**
     * Set this option to true to fetch the entire product catalog, false to limit the search to the products that your account has access to.
     */
    All?: boolean | null;
};