// src/api/huuray/models/ProductFromCatalogue.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ProductFromCatalogue = {
    ProductToken?: string | null;
    BrandName?: string | null;
    Country?: string | null;
    CountryCode?: string | null;
    Discount?: number | null;
    Denominations?: string | null;
    Currency?: string | null;
    RealTimeStock?: string | null;
    Categories?: string | null;
    LanguageCode?: string | null;
    BrandDescription?: string | null;
    RedemptionInstructions?: string | null;
    LogoFile?: string | null;
    isActive?: boolean | null;
    ecardId?: boolean | null;
};